Move ClosedRange helpers from MovingHeadVisualizer to baaahs.util.
